scripts/qlib/requirements.txt
# Qlib integration dependencies for My_Data vault
# Install with: pip install -r requirements.txt

# Microsoft Qlib — core quantitative research platform
pyqlib>=0.9.0

# Data manipulation — required by Qlib internals and custom scripts
pandas>=1.5.0

# Loads .env file so Python scripts share API keys with the Node.js pullers
python-dotenv>=1.0.0

# YAML parsing — reads Obsidian frontmatter from vault notes
pyyaml>=6.0

# LightGBM — default ML backend for Qlib alpha models
lightgbm>=3.3.0

# Matplotlib — chart generation for data pull output notes
matplotlib>=3.6.0

# yfinance — downloads fresh OHLCV data from Yahoo Finance for Qlib data bootstrap
yfinance>=0.2.0

# Rich — terminal UI for animated pipeline simulation (qlib sim command)
rich>=13.0
